requirements.txt
# Core Dependencies
streamlit>=1.28.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.10

# Vector Store & Embeddings
faiss-cpu
sentence-transformers

# PDF Processing
pypdf
pdf2image
pytesseract
pillow

# Utilities
python-dotenv
numpy
